See desc
Staticly typed almost everything. Unlike in Godot 4, this brings no performance improvement, but it brings peace of mind for debugging and improves readability.

Also set some variables to constants, and in the case of packed scenes, preloaded them as constants.

Moved the threaded loading into the Global singleton in an attempt to start background loading earlier on.

Used the above to make reloads after dying instant, utilizing the fact that the scene is already loaded in.